Added Makefile.in from the Waverous project, removed pgperf
In order to compile to C++ there was a long list of changes needed for
the Makefile. Included in this: removal of the custom version of gperf
that shipped with LambdaMOO since its earliest days. See the comment
in Makefile.in about the original purpose of the customized pgperf and
why it's no longer needed.