Fix void pointer problems when BYTECODE_REDUCE_REF is defined
When BYTECODE_REDUCE_REF is defined, a block of code is called that
wants to call the C standard library's qsort(). qsort() takes a
function pointer whose signature is:

     void
     qsort(void *base, size_t nel, size_t width,
         int (*compar)(const void *, const void *));

I found the solution on cplusplus.com, which is to cast the function
arguments within the body of the "compar" function.