Cleaned up some compiler errors missed by patching
Some casts for mymalloc() were missing, a couple instances of 'new'
needed to be '_new', and push_expr() was being fed a cast to Stmt
instead of Expr, probably human error on my part. When I did a lot of
the casting in the Waverous project I wrote Perl scripts to
mass-change the code base to avoid just such human errors.