Uncast arg to setjump(); casts for mymalloc()
This set of changes do not come from a patch made from a Waverous
commit; rather they are me playing catch-up to the most recent version
of LambdaMOO off of github, from which Stunt was forked.

Code using the Stream type is wrapped in a macro I had not seen
before, TRY_STREAM and ENDTRY_STREAM. I had to make a couple of
changes to the TRY and EXCEPT macros, which TRY_STREAM and
ENDTRY_STREAM use, removing a cast to void * on the first argument to
setjmp(). The struct used in those macros had a field "finally" that
needed an underscore prefix.